# Use the most significant bits for partitioning in the HashAggregate